+++ release_notes.md
@@ -1,4 +1,5 @@
-# Azure Functions CLI 4.2.3
+# Azure Functions CLI 4.3.0
+
 
 #### Host Version
 
@@ -6,5 +7,4 @@
 - In-Proc Host Version: 4.41.100 (4.841.100, 4.641.100)
 
 #### Changes
-
-- Updated assemblies to sign: AppService, Functions, WebJobs; consolidated OpenTelemetry binaries; removed Python exec patterns (t32, t64, w32, w64) (#4621)
+- Add `func pack` basic functionality (#4600)